added a component for rolesAppBar logic
